Fixed Heart and GemCounter GUI
So i fixed the heart and gemCounter gui thing, so now it doesn't only work for level one. I haven't implemented it into every level yet because I want to get the game over screen working so that I can just replace all the camera 2d's with the gui instead.