Added sounds and animation to the existing trampoline
so the main change was that the trampoline was adjusted. Thanks to sid it already existed and i just added some animations and sounds to it.
I also changed it to where the players are in FRONT of the portals instead of behind em. It was just a small detail that was bothering me.
Also i fixed the sound bus issue with the control menu, just a simple click of a button fixes it.